move viewoptions to option to keep paths simpler
